refactor(android): extract PlannedItemApi/Repository, clean suppress annotations
Split planned-item CRUD out of TodayActionsApi/TodayRepository into
dedicated PlannedItemApi + PlannedItemRepository, wired via Hilt.
Extract 5 Today card composables from HomeRoute into TodayItemCards.kt.
Remove @Suppress("TooManyFunctions") from TodayActionsApi (8 methods),
TodayRepository (11 methods), and HomeRoute (9 composables).
Drop redundant TooGenericExceptionCaught suppresses on passthrough methods
that delegate to the shared safeApiCall and never catch themselves.
Remove redundant per-function FunctionNaming suppresses covered by file-level.
