+++ android/app/src/main/java/com/daynest/android/data/today/TodayRepository.kt
@@ -3,6 +3,7 @@ package com.daynest.android.data.today
 import com.daynest.android.core.database.today.TodaySummaryDao
 import com.daynest.android.core.database.today.TodaySummaryEntity
 import com.daynest.android.core.model.TodaySummary
+import com.daynest.android.data.safeApiCall
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -12,7 +13,6 @@ import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-@Suppress("TooManyFunctions")
 class TodayRepository
     @Inject
     constructor(
@@ -48,15 +48,14 @@ class TodayRepository
                 Result.failure(e)
             }
 
-        @Suppress("TooGenericExceptionCaught")
         suspend fun completeChore(choreInstanceId: Int): Result<ChoreMutationDto> =
             safeApiCall { todayActionsApi.completeChore(choreInstanceId) }
 
-        @Suppress("TooGenericExceptionCaught", "ktlint:standard:function-signature")
+        @Suppress("ktlint:standard:function-signature")
         suspend fun skipChore(choreInstanceId: Int): Result<ChoreMutationDto> =
             safeApiCall { todayActionsApi.skipChore(choreInstanceId) }
 
-        @Suppress("TooGenericExceptionCaught", "ktlint:standard:function-signature")
+        @Suppress("ktlint:standard:function-signature")
         suspend fun rescheduleChore(
             choreInstanceId: Int,
             scheduledDate: String,
@@ -65,82 +64,30 @@ class TodayRepository
                 todayActionsApi.rescheduleChore(choreInstanceId, RescheduleChoreDto(scheduledDate = scheduledDate))
             }
 
-        @Suppress("TooGenericExceptionCaught")
         suspend fun completeTask(taskInstanceId: Int): Result<TaskMutationDto> =
             safeApiCall { todayActionsApi.completeTask(taskInstanceId) }
 
-        @Suppress("TooGenericExceptionCaught", "ktlint:standard:function-signature")
+        @Suppress("ktlint:standard:function-signature")
         suspend fun startTask(taskInstanceId: Int): Result<TaskMutationDto> =
             safeApiCall { todayActionsApi.startTask(taskInstanceId) }
 
-        @Suppress("TooGenericExceptionCaught", "ktlint:standard:function-signature")
+        @Suppress("ktlint:standard:function-signature")
         suspend fun skipTask(taskInstanceId: Int): Result<TaskMutationDto> =
             safeApiCall { todayActionsApi.skipTask(taskInstanceId) }
 
-        @Suppress("TooGenericExceptionCaught", "ktlint:standard:function-signature")
+        @Suppress("ktlint:standard:function-signature")
         suspend fun takeDose(doseInstanceId: Int): Result<DoseMutationDto> =
             safeApiCall { todayActionsApi.takeDose(doseInstanceId) }
 
-        @Suppress("TooGenericExceptionCaught", "ktlint:standard:function-signature")
+        @Suppress("ktlint:standard:function-signature")
         suspend fun skipDose(doseInstanceId: Int): Result<DoseMutationDto> =
             safeApiCall { todayActionsApi.skipDose(doseInstanceId) }
-
-        @Suppress("TooGenericExceptionCaught")
-        suspend fun markPlannedDone(
-            id: Int,
-            item: PlannedTodayItemDto,
-            isDone: Boolean,
-        ): Result<PlannedTodayItemDto> =
-            safeApiCall {
-                todayActionsApi.updatePlannedItem(
-                    id,
-                    PlannedItemUpdateDto(
-                        title = item.title,
-                        plannedFor = item.plannedFor,
-                        isDone = isDone,
-                        notes = item.notes,
-                        moduleKey = item.moduleKey,
-                        recurrenceHint = item.recurrenceHint,
-                        linkedSource = item.linkedSource,
-                        linkedRef = item.linkedRef,
-                    ),
-                )
-            }
-
-        @Suppress("TooGenericExceptionCaught")
-        suspend fun updatePlannedItem(
-            id: Int,
-            input: PlannedItemUpdateDto,
-        ): Result<PlannedTodayItemDto> = safeApiCall { todayActionsApi.updatePlannedItem(id, input) }
-
-        @Suppress("TooGenericExceptionCaught")
-        suspend fun deletePlannedItem(id: Int): Result<Unit> = safeApiCall { todayActionsApi.deletePlannedItem(id) }
-
-        @Suppress("TooGenericExceptionCaught")
-        suspend fun createPlannedItem(request: PlannedItemCreateDto): Result<PlannedTodayItemDto> =
-            safeApiCall { todayActionsApi.createPlannedItem(request) }
-
-        @Suppress("TooGenericExceptionCaught")
-        suspend fun listPlannedItems(
-            startDate: String?,
-            endDate: String?,
-        ): Result<List<PlannedTodayItemDto>> = safeApiCall { todayActionsApi.listPlannedItems(startDate, endDate) }
-
-        private fun TodaySummaryEntity.toDomain() =
-            TodaySummary(
-                routinesCount = routinesCount,
-                choresCount = choresCount,
-                medicationsCount = medicationsCount,
-                plannedPendingCount = plannedPendingCount,
-            )
-
-        @Suppress("TooGenericExceptionCaught")
-        private suspend fun <T> safeApiCall(call: suspend () -> T): Result<T> =
-            try {
-                Result.success(call())
-            } catch (e: CancellationException) {
-                throw e
-            } catch (e: Exception) {
-                Result.failure(e)
-            }
     }
+
+private fun TodaySummaryEntity.toDomain() =
+    TodaySummary(
+        routinesCount = routinesCount,
+        choresCount = choresCount,
+        medicationsCount = medicationsCount,
+        plannedPendingCount = plannedPendingCount,
+    )
